+++ src/common/custom-types.ts
@@ -163,6 +163,9 @@ interface CCBaseEditor extends BaseEditor {
   data?: Record<string, boolean | number | string>;
   // plugins can store values in maps under the plugin name
   plugins: Record<string, Record<string, any>>;
+
+  globalStyle: (style: React.CSSProperties) => React.CSSProperties;
+
   isMarkActive: (format: string) => boolean;
   toggleMark: (format: string, value?: any) => void;
   toggleSuperSubscript: (format: EFormat.subscript | EFormat.superscript) => void;
@@ -172,11 +175,7 @@ interface CCBaseEditor extends BaseEditor {
   selectedElements: () => CustomElement[];
 
   isElementEnabled: (format: string) => boolean;
-  configureElement: (format: string, controller: IDialogController, elt?: CustomElement) => void;
-
-  emitEvent: (event: string, ...args: any[]) => void;
-  onEvent: (event: string, handler: (...args: any[]) => void) => void;
-  offEvent: (event: string, handler: (...args: any[]) => void) => void;
+  configureElement: (format: string, controller?: IDialogController, elt?: CustomElement) => void;
 }
 
 export type CustomEditor = CCBaseEditor & ReactEditor & HistoryEditor;