+++ src/create-editor.ts
@@ -8,7 +8,7 @@ import { withColorMark } from "./plugins/color-plugin";
 import { withCoreBlocks } from "./plugins/core-blocks-plugin";
 import { withCoreInlines } from "./plugins/core-inlines-plugin";
 import { withCoreMarks } from "./plugins/core-marks-plugin";
-import { withEmitter } from "./plugins/emitter-plugin";
+import { withFontSize } from "./plugins/font-size-plugin";
 import { withImages } from "./plugins/image-plugin";
 import { withLinkInline } from "./plugins/link-plugin";
 import { withListBlocks } from "./plugins/list-plugin";
@@ -24,6 +24,8 @@ export function createEditor(options?: ICreateEditorOptions) {
 
   editor.plugins = {};
 
+  editor.globalStyle = () => ({});
+
   editor.isMarkActive = isMarkActive.bind(editor, editor);
   editor.toggleMark = (format: string, value: any = true) => toggleMark(editor, format, value);
   editor.toggleSuperSubscript = toggleSuperSubscript.bind(editor, editor);
@@ -35,8 +37,6 @@ export function createEditor(options?: ICreateEditorOptions) {
   editor.isElementEnabled = () => true;
   editor.configureElement = () => null;
 
-  editor = withEmitter(editor);
-
   editor = withCoreMarks(editor);
   editor = withColorMark(editor);
   editor = withCoreBlocks(editor);
@@ -44,6 +44,7 @@ export function createEditor(options?: ICreateEditorOptions) {
   editor = withCoreInlines(editor);
   editor = withImages(editor);
   editor = withLinkInline(editor);
+  editor = withFontSize(editor);
 
   // allow clients to attach their own plugins, etc.
   return onInitEditor?.(editor) ?? editor;