+++ src/editor-toolbar/editor-toolbar.stories.tsx
@@ -1,13 +1,15 @@
 import React from "react";
-import { EditorToolbar, getPlatformTooltip, IButtonSpec } from "./editor-toolbar";
 import IconBold from "../assets/icon-bold";
 import IconCode from "../assets/icon-code";
 import IconItalic from "../assets/icon-italic";
 import IconBulletedList from "../assets/icon-list-bulleted";
 import IconNumberedList from "../assets/icon-list-numbered";
 import IconUnderline from "../assets/icon-underline";
-import "./editor-toolbar.scss";
 import { EFormat } from "../common/slate-types";
+import { getPlatformTooltip, IButtonSpec } from "../common/toolbar-utils";
+import { EditorToolbar } from "./editor-toolbar";
+
+import "./editor-toolbar.scss";
 
 export default {
   title: "EditorToolbar"
@@ -19,42 +21,42 @@ const buttons: IButtonSpec[] = [
     format: EFormat.bold,
     SvgIcon: IconBold,
     tooltip: getPlatformTooltip("bold (mod-b)"),
-    isActive: false,
+    isActive: () => false,
     onClick: () => console.log("Bold")
   },
   {
     format: EFormat.italic,
     SvgIcon: IconItalic,
     tooltip: getPlatformTooltip("italic (mod-i)"),
-    isActive: true,
+    isActive: () => true,
     onClick: () => console.log("Italic")
   },
   {
     format: EFormat.underlined,
     SvgIcon: IconUnderline,
     tooltip: getPlatformTooltip("underline (mod-u)"),
-    isActive: false,
+    isActive: () => false,
     onClick: () => console.log("Underline")
   },
   {
     format: EFormat.code,
     SvgIcon: IconCode,
     tooltip: getPlatformTooltip("code"),
-    isActive: true,
+    isActive: () => true,
     onClick: () => console.log("Code")
   },
   {
     format: EFormat.numberedList,
     SvgIcon: IconNumberedList,
     tooltip: getPlatformTooltip("numbered list"),
-    isActive: false,
+    isActive: () => false,
     onClick: () => console.log("# List")
   },
   {
     format: EFormat.bulletedList,
     SvgIcon: IconBulletedList,
     tooltip: getPlatformTooltip("bulleted list"),
-    isActive: true,
+    isActive: () => true,
     onClick: () => console.log("• List")
   }
 ];